requirements.txt
python-magic==0.4.27
pyyaml==6.0
colorlog==6.7.0
mergedeep==1.3.4